+++ app/Models/UsersModel.php
@@ -9,4 +9,6 @@ class UsersModel extends Model
     protected $table = 'users';
     protected $primaryKey = 'id';
     protected $allowedFields = ['name','email','password'];
+    protected $useSoftDeletes = true;
+    protected $useTimestamps = true;
 }