+++ docs/changelog.md
@@ -1,3 +1,6 @@
+##### v0.0.14
+- Create Migration users
+ Create View admin/login
 ##### v0.0.13
 - Create Migration users
 - Create Controller Admin/users